finish Assignments Along The Layers move DTO from api to application rest with assignments $ submion

+++ Application/UseCases/Submission/DeleteSubmission/DeleteSubmissionHandler.cs
@@ -2,33 +2,40 @@
 using SmartGrader.Application.Common.Exceptions;
 using SmartGrader.Domain.Abstractions;
 using SmartGrader.Domain.Entities;
-using System;
-using System.Collections.Generic;
-using System.Linq;
-using System.Text;
-using System.Threading.Tasks;
 
 namespace SmartGrader.Application.UseCases.Submissions.DeleteSubmission
 {
-    public class DeleteSubmissionHandler : IRequestHandler<DeleteSubmissionCommand, Unit>
+    public class DeleteSubmissionHandler : IRequestHandler<DeleteSubmissionCommand>
     {
         private readonly ISubmissionRepository _repository;
         private readonly IUnitOfWork _unitOfWork;
 
-        public DeleteSubmissionHandler(ISubmissionRepository repository, IUnitOfWork unitOfWork)
-            => (_repository, _unitOfWork) = (repository, unitOfWork);
-
-        public async Task<Unit> Handle(DeleteSubmissionCommand request, CancellationToken cancellationToken)
+        public DeleteSubmissionHandler(
+            ISubmissionRepository repository,
+            IUnitOfWork unitOfWork)
         {
-            var submission = await _repository.GetByIdAsync(request.Id, cancellationToken);
+            _repository = repository;
+            _unitOfWork = unitOfWork;
+        }
+
+        public async Task Handle(
+            DeleteSubmissionCommand request,
+            CancellationToken cancellationToken)
+        {
+            var submission = await _repository.GetByIdAsync(
+                request.SubmissionId,
+                cancellationToken);
 
             if (submission is null)
-                throw new NotFoundException(nameof(Submission), request.Id);
+                throw new NotFoundException(nameof(Submission), request.SubmissionId);
+
+            if (submission.StudentId != request.StudentId)
+                throw new NotFoundException(
+                    "Submission does not belong to this student.",
+                    request.SubmissionId);
 
             await _repository.DeleteAsync(submission, cancellationToken);
             await _unitOfWork.SaveChangesAsync(cancellationToken);
-
-            return Unit.Value;
         }
     }
 }
